+++ packages/orq-rc/src/models/components/evaluatorresponseragas.ts
@@ -221,8 +221,8 @@ export const EvaluatorResponseRagas$inboundSchema: z.ZodType<
 > = z.object({
   _id: z.string(),
   description: z.string(),
-  created: z.string().default("2026-06-30T06:05:21.250Z"),
-  updated: z.string().default("2026-06-30T06:05:21.250Z"),
+  created: z.string().default("2026-06-30T07:21:18.786Z"),
+  updated: z.string().default("2026-06-30T07:21:18.786Z"),
   updated_by_id: z.nullable(z.string()).optional(),
   guardrail_config: z.nullable(
     z.union([