+++ packages/orq-rc/src/models/operations/postv2feedbackevaluation.ts
@@ -193,7 +193,7 @@ export const RequestBody3$outboundSchema: z.ZodType<
   source: PostV2FeedbackEvaluationRequestBodyFeedbackSource$outboundSchema
     .default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-06-30T06:05:40.468Z"))
+  reviewedAt: z.date().default(() => new Date("2026-06-30T07:21:49.931Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string_array"),
   values: z.array(z.string()),
@@ -247,7 +247,7 @@ export const PostV2FeedbackEvaluationRequestBody2$outboundSchema: z.ZodType<
     "orq",
   ),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-06-30T06:05:40.468Z"))
+  reviewedAt: z.date().default(() => new Date("2026-06-30T07:21:49.930Z"))
     .transform(v => v.toISOString()),
   type: z.literal("number"),
   value: z.number(),
@@ -305,7 +305,7 @@ export const PostV2FeedbackEvaluationRequestBody1$outboundSchema: z.ZodType<
   humanReviewId: z.string().optional(),
   source: RequestBodySource$outboundSchema.default("orq"),
   explanation: z.string().optional(),
-  reviewedAt: z.date().default(() => new Date("2026-06-30T06:05:40.467Z"))
+  reviewedAt: z.date().default(() => new Date("2026-06-30T07:21:49.930Z"))
     .transform(v => v.toISOString()),
   type: z.literal("string"),
   value: z.string(),