+++ packages/orq-rc/RELEASES.md
@@ -10369,3 +10369,13 @@ Based on:
 - [typescript v4.11.0-rc.58] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.58] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.58 - packages/orq-rc
+
+## 2026-06-30 07:23:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.59] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.59 - packages/orq-rc